+++ src/main/java/se/sundsvall/teamssender/auth/integration/DatabaseTokenCache.java
@@ -1,8 +1,10 @@
-package se.sundsvall.teamssender.auth.pojo;
+package se.sundsvall.teamssender.auth.integration;
 
 import com.microsoft.aad.msal4j.ITokenCacheAccessAspect;
 import com.microsoft.aad.msal4j.ITokenCacheAccessContext;
 import java.nio.charset.StandardCharsets;
+
+import se.sundsvall.teamssender.auth.model.TokenCacheEntity;
 import se.sundsvall.teamssender.auth.repository.ITokenCacheRepository;
 
 public class DatabaseTokenCache implements ITokenCacheAccessAspect {
@@ -17,10 +19,12 @@ public class DatabaseTokenCache implements ITokenCacheAccessAspect {
 
 	@Override
 	public void beforeCacheAccess(ITokenCacheAccessContext context) {
-		tokenCacheRepository.load(userId).ifPresent(cacheData -> {
-			String cacheString = new String(cacheData, StandardCharsets.UTF_8);
-			context.tokenCache().deserialize(cacheString);
-		});
+		tokenCacheRepository.findById(userId)
+				.map(TokenCacheEntity::getCacheData)
+				.ifPresent(cacheData -> {
+					String cacheString = new String(cacheData, StandardCharsets.UTF_8);
+					context.tokenCache().deserialize(cacheString);
+				});
 	}
 
 	@Override
@@ -28,7 +32,7 @@ public class DatabaseTokenCache implements ITokenCacheAccessAspect {
 		if (context.hasCacheChanged()) {
 			String serializedCache = context.tokenCache().serialize();
 			byte[] cacheData = serializedCache.getBytes(StandardCharsets.UTF_8);
-			tokenCacheRepository.save(userId, cacheData);
+			tokenCacheRepository.save(new TokenCacheEntity(userId, cacheData));
 		}
 	}
 }
