--- src/main/java/se/sundsvall/teamssender/auth/repository/TokenCacheJpaRepository.java
@@ -1,7 +1,0 @@
-package se.sundsvall.teamssender.auth.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import se.sundsvall.teamssender.auth.model.TokenCacheEntity;
-
-public interface TokenCacheJpaRepository extends JpaRepository<TokenCacheEntity, String> {
-}
