--- src/main/java/se/sundsvall/teamssender/auth/repository/TokenCacheRepository.java
@@ -1,25 +1,0 @@
-package se.sundsvall.teamssender.auth.repository;
-
-import java.util.Optional;
-import org.springframework.stereotype.Repository;
-import se.sundsvall.teamssender.auth.model.TokenCacheEntity;
-
-@Repository
-public class TokenCacheRepository implements ITokenCacheRepository {
-
-	private final TokenCacheJpaRepository tokenCacheJpaRepository;
-
-	public TokenCacheRepository(TokenCacheJpaRepository tokenCacheJpaRepository) {
-		this.tokenCacheJpaRepository = tokenCacheJpaRepository;
-	}
-
-	@Override
-	public Optional<byte[]> load(String userId) {
-		return tokenCacheJpaRepository.findById(userId).map(TokenCacheEntity::getCacheData);
-	}
-
-	@Override
-	public void save(String userId, byte[] cacheData) {
-		tokenCacheJpaRepository.save(new TokenCacheEntity(userId, cacheData));
-	}
-}
